refactor: domain layer 구성
