+++ lib/features/route_planning/data/repositories/route_repository_impl.dart
@@ -4,39 +4,7 @@ import 'package:ridingmate/features/route_planning/data/exceptions/route_excepti
 import 'package:ridingmate/features/route_planning/data/mappers/route_mapper.dart';
 import 'package:ridingmate/features/route_planning/data/models/route_api_response_model.dart';
 import 'package:ridingmate/features/route_planning/domain/entities/route_data.dart';
-
-enum RouteMode {
-  drivingCar,
-  cyclingRegular,
-  cyclingRoad,
-  cyclingMountain,
-  cyclingElectric,
-}
-
-extension RouteModeExtension on RouteMode {
-  String get apiValue {
-    switch (this) {
-      case RouteMode.drivingCar:
-        return 'driving-car';
-      case RouteMode.cyclingRegular:
-        return 'cycling-regular';
-      case RouteMode.cyclingRoad:
-        return 'cycling-road';
-      case RouteMode.cyclingMountain:
-        return 'cycling-mountain';
-      case RouteMode.cyclingElectric:
-        return 'cycling-electric';
-    }
-  }
-}
-
-abstract class RouteRepository {
-  Future<RouteData?> getRoute(
-    LatLng start,
-    LatLng end, {
-    RouteMode mode = RouteMode.cyclingRoad,
-  });
-}
+import 'package:ridingmate/features/route_planning/domain/repositories/route_repository.dart';
 
 class RouteRepositoryImpl implements RouteRepository {
   RouteRepositoryImpl({required RouteRemoteDataSource remoteDataSource})
